--- utils.py
@@ -1,219 +1,0 @@
-from _ctypes import PyObj_FromPtr
-import json
-import re
-import sys
-import time
-
-import numpy as np
-
-class NoIndent(object): # value wrapper
-    def __init__(self, value):
-        self.value = value
-
-class MyJSONEncoder(json.JSONEncoder): # see https://stackoverflow.com/questions/13249415/how-to-implement-custom-indentation-when-pretty-printing-with-the-json-module
-    FORMAT_SPEC = '@@{}@@'
-    regex = re.compile(FORMAT_SPEC.format(r'(\d+)'))
-
-    def __init__(self, **kwargs):
-        # save copy of any keyword argument values needed for use here
-        self.__sort_keys = kwargs.get('sort_keys', None)
-        super(MyJSONEncoder, self).__init__(**kwargs)
-
-    def default(self, obj):
-        return (self.FORMAT_SPEC.format(id(obj)) if isinstance(obj, NoIndent)
-                else super(MyJSONEncoder, self).default(obj))
-
-    def encode(self, obj):
-        format_spec = self.FORMAT_SPEC
-        json_repr = super(MyJSONEncoder, self).encode(obj) # default JSON repr
-
-        # replace any marked-up object ids in the JSON repr with the value returned from the json.dumps() of the corresponding wrapped object
-        for match in self.regex.finditer(json_repr):
-            id = int(match.group(1))
-            no_indent = PyObj_FromPtr(id)
-            json_obj_repr = json.dumps(no_indent.value, sort_keys=self.__sort_keys)
-
-            # replace the matched id string with json formatted representation of the corresponding object
-            json_repr = json_repr.replace(
-                            '"{}"'.format(format_spec.format(id)), json_obj_repr)
-        return json_repr
-
-class MyProgbar(object):
-    ''' Displays a progress bar.
-
-    Parameters
-    ----------
-    target: int 
-        Total number of steps expected, None if unknown.
-        
-    width: int 
-        Progress bar width on screen.
-    
-    verbose: int
-        Verbosity mode, 0 (silent), 1 (verbose), 2 (semi-verbose)
-    
-    stateful_metrics: list
-        Iterable of string names of metrics that should *not* be averaged over time. 
-        Metrics in this list will be displayed as-is. 
-        All others will be averaged by the progbar before display.
-
-    interval: float 
-        Minimum visual progress update interval (in seconds).
-    
-    unit_name: str
-        Display name for step counts (usually "step" or "sample").
-    '''
-
-    def __init__(self, target, width=30, verbose=1, interval=0.05, stateful_metrics=None, unit_name='step'):
-        self.target = target
-        self.width = width
-        self.verbose = verbose
-        self.interval = interval
-        self.unit_name = unit_name
-        if stateful_metrics:
-            self.stateful_metrics = set(stateful_metrics)
-        else:
-            self.stateful_metrics = set()
-        self._dynamic_display = ((hasattr(sys.stdout, 'isatty') and sys.stdout.isatty()) or 'ipykernel' in sys.modules or 'posix' in sys.modules)
-        self._total_width = 0
-        self._seen_so_far = 0
-        self._values = {}
-        self._values_order = []
-        self._start = time.time()
-        self._last_update = 0
-
-    def update(self, current, values=None):
-        ''' Updates the progress bar.
-
-        Parameters
-        ----------
-        current: int
-            Index of current step.
-        
-        values: list of tuples (name, value_for_last_step)
-            If 'name' is in 'stateful_metrics', 'value_for_last_step' will be displayed as-is.
-            Otherwise, an average of the metric over time will be displayed.
-        '''
-
-        values = values or []
-        for k, v in values:
-            if k not in self._values_order:
-                self._values_order.append(k)
-            if k not in self.stateful_metrics:
-                if k not in self._values:
-                    self._values[k] = [v * (current - self._seen_so_far), current - self._seen_so_far]
-                else:
-                    self._values[k][0] += v * (current - self._seen_so_far)
-                    self._values[k][1] += (current - self._seen_so_far)
-            else:
-                self._values[k] = [v, 1]
-        self._seen_so_far = current
-
-        now = time.time()
-        info = ' - %.0fs' % (now - self._start)
-        if self.verbose == 1:
-            if (now - self._last_update < self.interval and
-                    self.target is not None and current < self.target):
-                return
-
-            prev_total_width = self._total_width
-            if self._dynamic_display:
-                sys.stdout.write('\b' * prev_total_width)
-                sys.stdout.write('\r')
-            else:
-                sys.stdout.write('\n')
-
-            if self.target is not None:
-                numdigits = int(np.log10(self.target)) + 1
-                bar = ('%' + str(numdigits) + 'd/%d [') % (current, self.target)
-                prog = float(current) / self.target
-                prog_width = int(self.width * prog)
-                if prog_width > 0:
-                    bar += ('=' * (prog_width - 1))
-                    if current < self.target:
-                        bar += '>'
-                    else:
-                        bar += '='
-                bar += ('.' * (self.width - prog_width))
-                bar += ']'
-            else:
-                bar = '%7d/Unknown' % current
-
-            self._total_width = len(bar)
-            sys.stdout.write(bar)
-
-            if current:
-                time_per_unit = (now - self._start) / current
-            else:
-                time_per_unit = 0
-            if self.target is not None and current < self.target:
-                eta = time_per_unit * (self.target - current)
-                if eta > 3600:
-                    eta_format = '%d:%02d:%02d' % (eta // 3600, (eta % 3600) // 60, eta % 60)
-                elif eta > 60:
-                    eta_format = '%d:%02d' % (eta // 60, eta % 60)
-                else:
-                    eta_format = '%ds' % eta
-
-                info = ' - ETA: %s' % eta_format
-            else:
-                if time_per_unit >= 1 or time_per_unit == 0:
-                    info += ' %.0fs/%s' % (time_per_unit, self.unit_name)
-                elif time_per_unit >= 1e-3:
-                    info += ' %.0fms/%s' % (time_per_unit * 1e3, self.unit_name)
-                else:
-                    info += ' %.0fus/%s' % (time_per_unit * 1e6, self.unit_name)
-
-            for k in self._values_order:
-                info += ' - %s:' % k
-                if isinstance(self._values[k], list):
-                    avg = np.mean(self._values[k][0] / max(1, self._values[k][1]))
-                    if abs(avg) > 1e-3:
-                        info += ' %.4f' % avg
-                    else:
-                        info += ' %.4e' % avg
-                else:
-                    info += ' %s' % self._values[k]
-
-            self._total_width += len(info)
-            if prev_total_width > self._total_width:
-                info += (' ' * (prev_total_width - self._total_width))
-
-            if self.target is not None and current >= self.target:
-                info += '\n'
-
-            sys.stdout.write(info)
-            sys.stdout.flush()
-
-        elif self.verbose == 2:
-            if self.target is not None and current >= self.target:
-                numdigits = int(np.log10(self.target)) + 1
-                count = ('%' + str(numdigits) + 'd/%d') % (current, self.target)
-                info = count + info
-                for k in self._values_order:
-                    info += ' - %s:' % k
-                    avg = np.mean(self._values[k][0] / max(1, self._values[k][1]))
-                    if avg > 1e-3:
-                        info += ' %.4f' % avg
-                    else:
-                        info += ' %.4e' % avg
-                info += '\n'
-
-                sys.stdout.write(info)
-                sys.stdout.flush()
-
-        self._last_update = now
-
-    def add(self, n, values=None):
-        ''' Updates the progress bar.
-
-        Parameters
-        ----------
-        n: int
-            Index to add to current step.
-        
-        values: list of tuples (name, value_for_last_step)
-            If 'name' is in 'stateful_metrics', 'value_for_last_step' will be displayed as-is.
-            Otherwise, an average of the metric over time will be displayed.
-        '''
-        self.update(self._seen_so_far + n, values)
